+++ yash-cli/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to `yash-cli` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0-beta.1] - Unreleased
+
+### Changed
+
+- The shell now enables blocking reads on the standard input if it is a terminal
+  or a pipe as [required by POSIX](https://pubs.opengroup.org/onlinepubs/9699919799.2018edition/utilities/sh.html#tag_20_117_06).
+
 ## [0.1.0-alpha.1] - 2024-04-13
 
 ### Added